Finished distinct Addresstable
Have my 5 starting tables for dashboard development. Will have to add more in the upcoming days but this is a good starting spot.

Decided to keep the best address/city as just address and city, made the originals LinkAddress and LinkCity, the slightly cleaned up ones for geocoding as InputAddress and InputCity. For the actual Address and City, if the geocoding score wasn't good enough then used the Input versions.