Tracking down area discrepancies
Found an issue with totalrentable not matching buildingarea from an issue with sum() and having the wrong na rm call. There is some other issues with data values not matching despite being the same as far as I can tell. The land values from the properties will need some review. I removed what land values are in the buildings report but that may be too crude a solution.